chore: Enable more compiler warnings (#351)
As noted in https://github.com/apache/arrow-nanoarrow/issues/224, we can
use a format check attribute to ensure that format strings/types match
and check that `ArrowErrorCode` is always checked.

While I was in it for compiler warnings, I enabled a few more that
recently caused a problem in the Arrow R package.

To minimize impact on existing usage, these are only enabled when
`NANOARROW_DEBUG` is defined (as it is for all CMake debug builds and in
our CI).

+++ r/src/buffer.h
@@ -72,11 +72,11 @@ static inline SEXP buffer_borrowed_xptr(const void* addr, int64_t size_bytes,
 static inline void buffer_borrowed_xptr_set_type(SEXP buffer_xptr,
                                                  enum ArrowBufferType buffer_type,
                                                  enum ArrowType buffer_data_type,
-                                                 int32_t element_size_bits) {
+                                                 int64_t element_size_bits) {
   SEXP buffer_types_sexp = PROTECT(Rf_allocVector(INTSXP, 3));
   INTEGER(buffer_types_sexp)[0] = buffer_type;
   INTEGER(buffer_types_sexp)[1] = buffer_data_type;
-  INTEGER(buffer_types_sexp)[2] = element_size_bits;
+  INTEGER(buffer_types_sexp)[2] = (int32_t)element_size_bits;
   R_SetExternalPtrTag(buffer_xptr, buffer_types_sexp);
   UNPROTECT(1);
 }
